feat!: extract plugins and packages as  separate packages
