+++ Cargo.toml
@@ -1,51 +1,42 @@
-[package]
-name = "ssufid"
-version = "0.1.0"
+[workspace]
+resolver = "3"
+members = ["packages/*", "plugins/*"]
+
+[workspace.package]
 edition = "2024"
 description = "SSU Announcement Crawler for Everyone"
 license = "MIT"
 authors = [
-    "Hyomin Koo <me@eatsteak.dev>",
-    "이중곤 <2wndrhs@gmail.com>",
-    "Jung Haesung <harryjung1001@gmail.com>",
+  "Hyomin Koo <me@eatsteak.dev>",
+  "이중곤 <2wndrhs@gmail.com>",
+  "Jung Haesung <harryjung1001@gmail.com>",
 ]
 
-[features]
-default = ['rss']
-rss = ['dep:rss']
-
-[dependencies]
-eyre = "0.6.12"
-reqwest = { version = "0.12.14", features = [
-    "charset",
-    "http2",
-    "macos-system-configuration",
-    "rustls-tls",
-    "cookies",
-    "gzip",
-    "brotli",
-] }
-scraper = "0.23.1"
+[workspace.dependencies]
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.140"
-thiserror = "2.0.12"
-time = { version = "0.3.41", features = [
-    "serde",
-    "macros",
-    "formatting",
-    "parsing",
+time = "0.3.41"
+reqwest = { version = "0.12.14", features = [
+  "charset",
+  "http2",
+  "macos-system-configuration",
+  "rustls-tls",
+  "cookies",
+  "gzip",
+  "brotli",
 ] }
-tokio = { version = "1.44.2", features = ["full"] }
-rss = { version = "2.0", optional = true }
+scraper = "0.23.1"
+thiserror = "2.0.12"
+tokio = { version = "1.45.0", features = ["full"] }
 url = "2.5.4"
 futures = "0.3.31"
-color-eyre = "0.6.3"
-clap = { version = "4.5.36", features = ["derive"] }
 log = "0.4.27"
-env_logger = "0.11.8"
-serde_yaml = "0.9.34"
 mime_guess = "2.0.5"
 
-[dev-dependencies]
-dotenv = "0.15.0"
-time = { version = "0.3.40", features = ["macros"] }
+ssufid = { path = "packages/ssufid", features = ["rss"] }
+
+ssufid_itsites = { path = "plugins/it_sites" }
+ssufid_media = { path = "plugins/media" }
+ssufid_mediamba = { path = "plugins/mediamba" }
+ssufid_ssucatch = { path = "plugins/ssucatch" }
+ssufid_ssupath = { path = "plugins/ssupath" }
