+++ packages/cli/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "ssufid_cli"
+version = "0.1.0"
+description.workspace = true
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+
+[dependencies]
+eyre = "0.6.12"
+color-eyre = "0.6.3"
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+time = { workspace = true, features = [
+  "serde",
+  "macros",
+  "formatting",
+  "parsing",
+] }
+futures = { workspace = true }
+tokio = { workspace = true, features = ["full"] }
+clap = { version = "4.5.36", features = ["derive"] }
+log = { workspace = true }
+env_logger = "0.11.8"
+
+ssufid = { workspace = true }
+ssufid_itsites = { workspace = true }
+ssufid_media = { workspace = true }
+ssufid_mediamba = { workspace = true }
+ssufid_ssucatch = { workspace = true }
+ssufid_ssupath = { workspace = true }
+
+[dev-dependencies]
+time = { version = "0.3.40", features = ["macros"] }
