+++ packages/ssufid/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "ssufid"
+version = "0.1.0"
+description.workspace = true
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+
+
+[features]
+default = ['rss']
+rss = ['dep:rss']
+
+[dependencies]
+reqwest = { workspace = true, features = [
+    "charset",
+    "http2",
+    "macos-system-configuration",
+    "rustls-tls",
+    "cookies",
+    "gzip",
+    "brotli",
+] }
+scraper = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+time = { workspace = true, features = [
+    "serde",
+    "macros",
+    "formatting",
+    "parsing",
+] }
+tokio = { workspace = true, features = ["full"] }
+rss = { version = "2.0", optional = true }
+url = { workspace = true }
+futures = { workspace = true }
+log = { workspace = true }
+mime_guess = { workspace = true }
+
+[dev-dependencies]
+time = { workspace = true, features = ["macros"] }
