+++ packages/ssufid/src/core/rss.rs
