+++ packages/ssufid/src/error.rs
