+++ packages/ssufid/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod core;
 pub mod error;
-pub mod plugins;
 
 pub use core::SsufidCore;
