+++ plugins/it_sites/src/cse/bachelor.rs
@@ -1,32 +1,32 @@
-use crate::{
+use crate::common::it_crawler::ItCrawler;
+use ssufid::{
     PluginError,
     core::{SsufidPlugin, SsufidPost},
 };
 
-use crate::plugins::common::it_crawler::ItCrawler;
-
-pub struct CseEmploymentPlugin {
+pub struct CseBachelorPlugin {
     crawler: ItCrawler<Self>,
 }
 
-impl SsufidPlugin for CseEmploymentPlugin {
-    const IDENTIFIER: &'static str = "cse.ssu.ac.kr/employment";
-    const TITLE: &'static str = "숭실대학교 컴퓨터학부 취업정보";
-    const DESCRIPTION: &'static str = "숭실대학교 컴퓨터학부 홈페이지의 취업정보를 제공합니다.";
-    const BASE_URL: &'static str = "https://cse.ssu.ac.kr/bbs/board.php?bo_table=employment";
+impl SsufidPlugin for CseBachelorPlugin {
+    const IDENTIFIER: &'static str = "cse.ssu.ac.kr/bachelor";
+    const TITLE: &'static str = "숭실대학교 컴퓨터학부 학사 공지사항";
+    const DESCRIPTION: &'static str =
+        "숭실대학교 컴퓨터학부 홈페이지의 학사 공지사항을 제공합니다.";
+    const BASE_URL: &'static str = "https://cse.ssu.ac.kr/bbs/board.php?bo_table=notice";
 
     async fn crawl(&self, posts_limit: u32) -> Result<Vec<SsufidPost>, PluginError> {
         self.crawler.crawl(posts_limit).await
     }
 }
 
-impl Default for CseEmploymentPlugin {
+impl Default for CseBachelorPlugin {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl CseEmploymentPlugin {
+impl CseBachelorPlugin {
     fn new() -> Self {
         Self {
             crawler: ItCrawler::new(),
@@ -41,7 +41,7 @@ mod tests {
     #[tokio::test]
     async fn test_crawl() {
         let posts_limit = 100;
-        let plugin = CseEmploymentPlugin::new();
+        let plugin = CseBachelorPlugin::new();
         let posts = plugin.crawl(posts_limit).await.unwrap();
         assert_eq!(posts.len(), posts_limit as usize);
         // println!("{:#?}", posts);
