+++ plugins/it_sites/src/cse/graduate.rs
@@ -1,32 +1,33 @@
-use crate::plugins::common::it_crawler::ItCrawler;
-use crate::{
+use ssufid::{
     PluginError,
     core::{SsufidPlugin, SsufidPost},
 };
 
-pub struct CseBachelorPlugin {
+use crate::common::it_crawler::ItCrawler;
+
+pub struct CseGraduatePlugin {
     crawler: ItCrawler<Self>,
 }
 
-impl SsufidPlugin for CseBachelorPlugin {
-    const IDENTIFIER: &'static str = "cse.ssu.ac.kr/bachelor";
-    const TITLE: &'static str = "숭실대학교 컴퓨터학부 학사 공지사항";
+impl SsufidPlugin for CseGraduatePlugin {
+    const IDENTIFIER: &'static str = "cse.ssu.ac.kr/graduate";
+    const TITLE: &'static str = "숭실대학교 컴퓨터학부 대학원 공지사항";
     const DESCRIPTION: &'static str =
-        "숭실대학교 컴퓨터학부 홈페이지의 학사 공지사항을 제공합니다.";
-    const BASE_URL: &'static str = "https://cse.ssu.ac.kr/bbs/board.php?bo_table=notice";
+        "숭실대학교 컴퓨터학부 홈페이지의 대학원 공지사항을 제공합니다.";
+    const BASE_URL: &'static str = "https://cse.ssu.ac.kr/bbs/board.php?bo_table=gra_notice";
 
     async fn crawl(&self, posts_limit: u32) -> Result<Vec<SsufidPost>, PluginError> {
         self.crawler.crawl(posts_limit).await
     }
 }
 
-impl Default for CseBachelorPlugin {
+impl Default for CseGraduatePlugin {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl CseBachelorPlugin {
+impl CseGraduatePlugin {
     fn new() -> Self {
         Self {
             crawler: ItCrawler::new(),
@@ -41,7 +42,7 @@ mod tests {
     #[tokio::test]
     async fn test_crawl() {
         let posts_limit = 100;
-        let plugin = CseBachelorPlugin::new();
+        let plugin = CseGraduatePlugin::new();
         let posts = plugin.crawl(posts_limit).await.unwrap();
         assert_eq!(posts.len(), posts_limit as usize);
         // println!("{:#?}", posts);
