+++ plugins/it_sites/src/sw/bachelor.rs
@@ -1,33 +1,32 @@
-use crate::{
+use crate::common::it_crawler::ItCrawler;
+use ssufid::{
     PluginError,
     core::{SsufidPlugin, SsufidPost},
 };
 
-use crate::plugins::common::it_crawler::ItCrawler;
-
-pub struct SwGraduatePlugin {
+pub struct SwBachelorPlugin {
     crawler: ItCrawler<Self>,
 }
 
-impl SsufidPlugin for SwGraduatePlugin {
-    const IDENTIFIER: &'static str = "sw.ssu.ac.kr/graduate";
-    const TITLE: &'static str = "숭실대학교 소프트웨어 대학원 공지사항";
+impl SsufidPlugin for SwBachelorPlugin {
+    const IDENTIFIER: &'static str = "sw.ssu.ac.kr/bachelor";
+    const TITLE: &'static str = "숭실대학교 소프트웨어학부 학사 공지사항";
     const DESCRIPTION: &'static str =
-        "숭실대학교 컴퓨터학부 홈페이지의 대학원 공지사항을 제공합니다.";
-    const BASE_URL: &'static str = "https://sw.ssu.ac.kr/bbs/board.php?bo_table=gra_notice";
+        "숭실대학교 소프트웨어 홈페이지의 학사 공지사항을 제공합니다.";
+    const BASE_URL: &'static str = "https://sw.ssu.ac.kr/bbs/board.php?bo_table=notice";
 
     async fn crawl(&self, posts_limit: u32) -> Result<Vec<SsufidPost>, PluginError> {
         self.crawler.crawl(posts_limit).await
     }
 }
 
-impl Default for SwGraduatePlugin {
+impl Default for SwBachelorPlugin {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl SwGraduatePlugin {
+impl SwBachelorPlugin {
     fn new() -> Self {
         Self {
             crawler: ItCrawler::new(),
@@ -42,7 +41,7 @@ mod tests {
     #[tokio::test]
     async fn test_crawl() {
         let posts_limit = 100;
-        let plugin = SwGraduatePlugin::new();
+        let plugin = SwBachelorPlugin::new();
         let posts = plugin.crawl(posts_limit).await.unwrap();
         assert_eq!(posts.len(), posts_limit as usize);
         // println!("{:#?}", posts);
