--- src/plugins/mod.rs
@@ -1,8 +1,0 @@
-pub mod cse;
-pub mod media;
-pub mod sec;
-pub mod ssu_catch;
-pub mod ssu_path;
-pub mod sw;
-
-mod common;
